Pulling some functions out into their own module and using browserify
