+++ package.json
@@ -8,7 +8,10 @@
   },
   "devDependencies": {
     "bower": "~>1.2.8",
+    "browserify": "~6.2.0",
     "require-dir": "~0.1.0",
+    "es5-shim": "~4.0.3",
+    "vinyl-source-stream": "~1.0.0",
     "gulp": "^3.5.6",
     "gulp-replace": "^0.2.0",
     "gulp-uglify": "^0.2.1",
@@ -23,7 +26,8 @@
     "gulp-s3": "nkostelnik/gulp-s3#af55865",
     "gulp-livereload": "^1.2.0",
     "gulp-connect": "^1.0.9",
-    "gulp-task-listing": "^0.2.2"
+    "gulp-task-listing": "^0.2.2",
+    "gulp-streamify": "~0.0.4"
   },
   "scripts": {
     "build": "gulp"
